treewide: split the data mod into its own crate
The test cases written for data were really meant for validating a beer.json
at change submission time (e.g. PR), so move the data structs to that repo
instead.

metric-beer-data is now the authority for the format, any changes to the
format must also be made to the schema (i.e. structs) and the data
simulataneously.

NOTE: The old javascript version of metric-beer does not reference the structs,
just the data. That version will have to be updated manually if there is a
breaking change to the format.